requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
colorama==0.4.6